pyproject.toml
[tool.poetry]
name = "continuing-education"
version = "0.1.0"
description = ""
authors = ["Ryan Peach <rgpeach10@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
gymnasium = "^0.29.1"
numpy = "^1.26.4"
pydantic = "^2.6.4"
torch = "^2.2.1"
pillow = "^10.2.0"
pygame = "^2.5.2"
gym-games = "^1.0.3"
MinAtar = "^1.0.15"
huggingface-hub = "^0.21.4"
ipytest = "^0.14.0"
ple = {path = "submodules/PyGame-Learning-Environment"}
plotly = "^5.20.0"
sympy = "^1.12"
tqdm = "^4.66.2"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pytest = "^8.1.1"
mypy = "^1.9.0"
jupyter = "^1.0.0"
jupytext = "^1.16.1"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-include = ["*.ipynb"]
exclude = ["./submodules"]

[tool.ruff.lint]
extend-ignore = ["E402"]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"

[[tool.mypy.overrides]]
module = "plotly.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ipytest.*"
ignore_missing_imports = true
